refactor main entry package name
