+++ src/adyen/api_test.go
@@ -4,7 +4,7 @@
  * Contact: support@adyen.com
  */
 
-package api
+package adyen
 
 import (
 	"os"
